software/requirements.txt
pyserial
hidapi
